CartApi and CartApiImpl were moved to api package
